[feat] Hero 섹션 구현
- 3초마다 메인 백그라운드 그림 변경
- dot 클릭 시, 해당 그림으로 변경
